requirements.txt
aiogram
python-dotenv
sqlalchemy
asyncpg
alembic
psycopg2-binary